+++ server/modules/app.module.ts
@@ -15,6 +15,7 @@ import { TestrunsModule } from './testruns/testruns.module';
 import { DevicesModule } from './devices/devices.module';
 import { AttachmentsModule } from './attachments/attachments.module';
 import { TestResultsModule } from './testresults/testresults.module';
+import { NotificationsModule } from './notifications/notifications.module';
 
 import { MongoDBConnectionStringBuilder } from './connection-string.builder';
 
@@ -25,8 +26,9 @@ import { MongoDBConnectionStringBuilder } from './connection-string.builder';
         .setUser(config.get('db.user'))
         .setPassword(config.get('db.password'))
         .setHost(config.get('db.host'))
+        .setPort(config.get('db.port'))
         .setDatabaseName(config.get('db.name'))
-        .build()),
+        .build(), { useNewUrlParser: true }),
     AuthModule,
     ProjectsModule,
     TestcasesModule,
@@ -36,7 +38,8 @@ import { MongoDBConnectionStringBuilder } from './connection-string.builder';
     RolesModule,
     DevicesModule,
     AttachmentsModule,
-    TestResultsModule
+    TestResultsModule,
+    NotificationsModule
   ],
   controllers: [],
   providers: [],