+++ src/app/test-management/testcase-review-notification/testcase-review-notification.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { TestcaseReviewNotificationComponent } from './testcase-review-notification.component';
+
+xdescribe('TestcaseReviewNotificationComponent', () => {
+  let component: TestcaseReviewNotificationComponent;
+  let fixture: ComponentFixture<TestcaseReviewNotificationComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ TestcaseReviewNotificationComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(TestcaseReviewNotificationComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});