+++ src/app/test-management/testcase-review/testcase-status.ts
@@ -1,0 +1,8 @@
+export enum TestcaseStatus {
+  CREATED,
+  APPROVED,
+  REJECTED,
+  AUTOMATION_CANDIDATE,
+  AUTOMATED,
+  ARCHIEVED
+}